# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisSummary": {
        "description": "An analysis without its report -- what a list endpoint returns.",
        "properties": {
          "analysis_id": {
            "title": "Analysis Id",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "default": "",
            "title": "Created At",
            "type": "string"
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/CriterionProgress"
            },
            "title": "Criteria",
            "type": "array"
          },
          "document_id": {
            "title": "Document Id",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "progress": {
            "$ref": "#/components/schemas/Progress"
          },
          "stage": {
            "default": "",
            "title": "Stage",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "done",
              "failed",
              "cancelled",
              "interrupted"
            ],
            "title": "Status",
            "type": "string"
          },
          "totals": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisTotals"
              },
              {
                "type": "null"
              }
            ]
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "analysis_id",
          "document_id",
          "status"
        ],
        "title": "AnalysisSummary",
        "type": "object"
      },
...
      "AnalyzeRequest": {
        "properties": {
          "criteria": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Criterion ids to run. Omit for all five.",
            "title": "Criteria"
          },
          "document_id": {
            "description": "The document to analyse. Required: there is no implicit current document.",
            "title": "Document Id",
            "type": "integer"
          }
        },
        "required": [
          "document_id"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
...
      "Error": {
        "description": "The body of every 4xx and 5xx this API returns.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "Error",
        "type": "object"
      },
...
      "Progress": {
        "properties": {
          "done": {
            "default": 0,
            "title": "Done",
            "type": "integer"
          },
          "total": {
            "default": 0,
            "title": "Total",
            "type": "integer"
          }
        },
        "title": "Progress",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "\nCompliance analysis of PDF contracts, and cited chat over them.\n\nUpload a contract to get a `document_id`, then bind everything else to it:\nanalysis and chat both require one, and retrieval is scoped to it in the\nlibrary rather than in this API, so an answer cannot cite another contract.\n\nAnalyses are jobs. `POST /analyses` returns an `analysis_id` in under a second\nand the run takes a minute or more; poll `GET /analyses/{analysis_id}` or\nsubscribe to its `/events`. Chat streams by default and can return one JSON\nbody instead.\n\nErrors are always `{\"error\": {\"code\", \"message\", \"hint\"}}`. `code` is stable\nand `hint` says what to do next, which is what a model driving this API through\nan MCP server needs in order to recover.\n",
    "title": "Contract Analyzer",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/analyses": {
      "get": {
        "description": "Live and stored analyses merged, so a run from a previous boot or from `make analyze` is listed too. `document_id` is required: a global list is a KPI question, and `/metrics/runs` is where it belongs.",
        "operationId": "index_analyses_get",
        "parameters": [
          {
            "in": "query",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AnalysisSummary"
                  },
                  "title": "Response Index Analyses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Malformed request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing or wrong X-API-Key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such document or analysis."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The resource is busy, or already in the requested state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The upload is over api_max_upload_mb."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Only PDF uploads are supported."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request body did not validate."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unhandled server error; the X-Trace-Id is on every log line."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "An upstream call failed after its retries."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A key or a dependency the operation needs is not configured."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "One document's analyses, newest first",
        "tags": [
          "analyses"
        ]
      },
      "post": {
        "description": "Returns immediately with an `analysis_id`; poll `GET /analyses/{id}` or subscribe to `GET /analyses/{id}/events`. A submission that matches one already queued or running returns that analysis with `200` instead of starting a second run -- send an `Idempotency-Key` header to force a new one.",
        "operationId": "submit_analyses_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Malformed request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing or wrong X-API-Key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such document or analysis."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The resource is busy, or already in the requested state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The upload is over api_max_upload_mb."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Only PDF uploads are supported."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request body did not validate."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unhandled server error; the X-Trace-Id is on every log line."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "An upstream call failed after its retries."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A key or a dependency the operation needs is not configured."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Queue an analysis of one contract",
        "tags": [
          "analyses"
        ]
      }
    },
    "/analyses/{analysis_id}": {
      "get": {
        "description": "Answered from the live job when this process is running it, and from the stored record otherwise -- a finished analysis returns its report after a restart. `detail=summary` omits quotes and rationale -- the MCP default, because a full report is a lot of context to put in a model's window.",
        "operationId": "detail_analyses__analysis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "detail",
            "required": false,
            "schema": {
              "default": "full",
              "enum": [
                "full",
                "summary"
              ],
              "title": "Detail",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analysis"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Malformed request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing or wrong X-API-Key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such document or analysis."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The resource is busy, or already in the requested state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The upload is over api_max_upload_mb."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Only PDF uploads are supported."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request body did not validate."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unhandled server error; the X-Trace-Id is on every log line."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "An upstream call failed after its retries."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A key or a dependency the operation needs is not configured."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "An analysis, with its report once it is done",
        "tags": [
          "analyses"
        ]
      }
    },
    "/analyses/{analysis_id}/cancel": {
      "post": {
        "description": "Criteria that have not started are skipped and listed in the report's `skipped`. One already talking to the model finishes: stopping it would mean interrupting the agent loop between tool calls, which this API does not do.",
        "operationId": "cancel_analyses__analysis_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Malformed request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing or wrong X-API-Key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such document or analysis."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The resource is busy, or already in the requested state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The upload is over api_max_upload_mb."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Only PDF uploads are supported."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request body did not validate."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unhandled server error; the X-Trace-Id is on every log line."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "An upstream call failed after its retries."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A key or a dependency the operation needs is not configured."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Stop an analysis after the criteria already in flight",
        "tags": [
          "analyses"
        ]
      }
    },
    "/analyses/{analysis_id}/events": {
      "get": {
        "description": "`status`, one `criterion` per verdict, `tool_call` for each search the agents make, then `done` or `error`, after which the stream closes. Subscribing late replays what was missed; subscribing after the job finished returns the replay and closes, rather than hanging.",
        "operationId": "events_analyses__analysis_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Malformed request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing or wrong X-API-Key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such document or analysis."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The resource is busy, or already in the requested state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The upload is over api_max_upload_mb."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Only PDF uploads are supported."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request body did not validate."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unhandled server error; the X-Trace-Id is on every log line."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "An upstream call failed after its retries."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A key or a dependency the operation needs is not configured."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Progress as server-sent events",
        "tags": [
          "analyses"
        ]
      }
    },
...
  },
...
}